+++ internal/season/repository.go
@@ -5,7 +5,7 @@ import "github.com/statistico/statistico-data/internal/model"
 type Repository interface {
 	Insert(s *model.Season) error
 	Update(s *model.Season) error
-	Id(id int) (*model.Season, error)
-	Ids() ([]int, error)
-	CurrentSeasonIds() ([]int, error)
+	Id(id int64) (*model.Season, error)
+	Ids() ([]int64, error)
+	CurrentSeasonIds() ([]int64, error)
 }
